requirements.txt
flask
flask-socketio
selenium
pyperclip
pyinstaller
python-socketio
python-engineio
eventlet
flask-debugtoolbar
webdriver-manager
ffmpeg-python